Keep query state out of overlay windows
